% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleHeatmap.R
\name{AddPal}
\alias{AddPal}
\title{AddPal}
\usage{
AddPal(annot_name, row_annot, col_annot)
}
\arguments{
\item{annot_name}{Character, name of annotation}

\item{row_annot}{Dataframe of row annotations}

\item{col_annot}{Dataframe of column annotations}
}
\value{
Named character vector of hex colours
}
\description{
Add palette, helper function for simpleHeatmap()
}
