% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colours.R
\name{blend_colours}
\alias{blend_colours}
\title{Blend colours}
\usage{
blend_colours(colour1, colour2, n)
}
\arguments{
\item{colour1}{A hex character string}

\item{colour2}{A hex character string}

\item{n}{Integer specifying number of colours}
}
\value{
Character vector of hex strings
}
\description{
Blend colours
}
