% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunc_R.R
\name{calc_pca}
\alias{calc_pca}
\title{calc_pca}
\usage{
calc_pca(data, is.log, centre)
}
\arguments{
\item{data}{data.frame or matrix object}

\item{is.log}{TRUE or FALSE are the data already log-transformed?}

\item{centre}{TRUE or FALSE should the data be centred by scale()}
}
\value{
List containing SVD from PCA calculation (sing.val) and percent variance (var)
}
\description{
Calculates PCA for input data
}
