% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scWrangle.R
\name{Check_obj}
\alias{Check_obj}
\title{Check_obj}
\usage{
Check_obj(dat_obj, desired_class, dat_assays = "raw")
}
\arguments{
\item{dat_obj}{SingleCellExperiment, SummarizedExperiment or Seurat object}

\item{desired_class}{Character string indicating desired class, one of c("Seurat", "SummarizedExperiment","SingleCellExperiment")}

\item{dat_assays}{Either "raw" or "logcounts" indicating the Seurat assay to use when converting Seurat objects to SingleCellExperiment}
}
\value{
Object of desired class
}
\description{
Checks the class of input object and converts if required
}
