% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleHeatmap.R
\name{CheckCols}
\alias{CheckCols}
\title{CheckCols}
\usage{
CheckCols(col_lst, row_annot, col_annot)
}
\arguments{
\item{col_lst}{List of columns}

\item{row_annot}{Dataframe of row annotations}

\item{col_annot}{Dataframe of column annotations}
}
\value{
List of columns
}
\description{
Helper function for simpleHeatmap()
}
