% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colours.R
\name{ColourBar}
\alias{ColourBar}
\title{ColourBar}
\usage{
ColourBar(min, max, n_ticks, col.vector, flip = TRUE)
}
\arguments{
\item{min}{Numeric, minimum value}

\item{max}{Numeric, maximum value}

\item{n_ticks}{Numeric, number of ticks to plot in colour bar}

\item{col.vector}{Character vector, colour bar will plot will contain same number of values as col.vector}

\item{flip}{Boolean, if TRUE flips coordinates}
}
\value{
A ggplot object that looks like a colour bar
}
\description{
Plot custom colour bar to use as a customisable legend for continuous plots
}
