% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleHeatmap.R
\name{dict2Annotdf}
\alias{dict2Annotdf}
\title{dict2Annotdf}
\usage{
dict2Annotdf(dict_lst, data_names)
}
\arguments{
\item{dict_lst}{Named list of dictionaries}

\item{data_names}{Vector of row or column names, ensures ordering}
}
\value{
Dataframe of annotations in the correct format for `simpleHeatmap`, where rownames are heatmap_dat rownames
}
\description{
Converts dictionaries to annotation dataframe. Helper function for simpleHeatmap().
}
\examples{
col_annot <- dict2Annotdf(list(SigType=sig_group, Author=SigAuthor), colnames(heatmap_dat))
row_annot <- dict2Annotdf(list(CellType=setNames(tmp_dge$samples$CellType,rownames(tmp_dge$samples)),
                               Cluster=setNames(tmp_dge$samples$Cluster,rownames(tmp_dge$samples)),
                               Stage= setNames(tmp_dge$samples$Stage, rownames(tmp_dge$samples))),
                                  rownames(heatmap_dat))
simpleHeatmap(heatmap_dat,row_annot,col_annot)
}
