% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunc_R.R
\name{ExtractLegends}
\alias{ExtractLegends}
\title{ExtractLegends}
\usage{
ExtractLegends(plot_x, n_row = 2, ...)
}
\arguments{
\item{plot_x}{ggplot object}

\item{n_row}{Integer, number of rows of keys in each legend}

\item{...}{}
}
\value{
A gtable object holding just the legend or NULL if there is no legend (from cowplot::get_legend())
}
\description{
Used by ggMulti() to extract legends and orient properly for cowplot::plot_grid()
}
\examples{

legend_list <- list(ExtractLegends(A),ExtractLegends(B))
ggMulti(list(A,B),leg_lst=legend_list,leg_nrow=1)

legend_list <- list(ExtractLegends(A),ExtractLegends(B))

}
