% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colours.R
\name{get_cbarVals}
\alias{get_cbarVals}
\title{get_cbarVals}
\usage{
get_cbarVals(x, n)
}
\arguments{
\item{x}{Numeric vector of values for a colour bar}

\item{n}{Number of output values to condense x into}
}
\value{
Numeric vector of length n containing new colour bar values
}
\description{
Scale values for a colour bar
}
