% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colours.R
\name{GetColPal}
\alias{GetColPal}
\title{GetColPal}
\usage{
GetColPal(vec)
}
\arguments{
\item{vec}{Character vector of names}
}
\value{
Named character vector containing a hex colour for each name in `vec`
}
\description{
Returns a named colour palette for the input character vector using the first option in GetColPal_options().
The purpose of this function is to obtain quick or default colour palettes that are prettier than ggplot defaults
}
