% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colours.R
\name{GetColPal_options}
\alias{GetColPal_options}
\title{GetColPal_options}
\usage{
GetColPal_options(n)
}
\arguments{
\item{n}{Integer, length of desired colour pal}
}
\value{
List of colour paletes for length `n`
}
\description{
Returns colour palette options of a given length, n.
}
\details{
This function has curated palettes from a variety of other colour packages such as wesanderson, Manu, and ggsci
}
