% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunc_R.R
\name{GetLegend}
\alias{GetLegend}
\title{GetLegend}
\usage{
GetLegend(
  col_dict,
  lab_str,
  title_pos = "top",
  n_row = 1,
  by_row = TRUE,
  rl = 1,
  leg_size = 4,
  buffer = margin(0, 0, 0, 0),
  ...
)
}
\arguments{
\item{col_dict}{Named character vector, where values are hex colours}

\item{lab_str}{Character string, title of legend}

\item{title_pos}{c("top", "bottom") for vertical, c("left", "right") for horiztonal}

\item{n_row}{Integer, number of rows}

\item{by_row}{TRUE or FALSE}

\item{rl}{Relative text size}

\item{buffer}{Legend margin i.e. margin(t,r,b,l)}
}
\value{
A gtable object holding just the legend (from cowplot::get_legend())
}
\description{
Returns a gtable containing a custom colour legend based on col_dict
}
