% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleHeatmap.R
\name{GetOrdering}
\alias{GetOrdering}
\title{GetOrdering}
\usage{
GetOrdering(annot_df, orderBy)
}
\arguments{
\item{annot_df}{Dataframe of annotations like `dict2Annotdf`}

\item{orderBy}{Character, column of `annot_df` that should be ordered in heatmap}
}
\value{
Annotation dataframe, ordered according to `orderBy`
}
\description{
Helper function forsimpleHeatmap()
}
