% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scWrangle.R
\name{GetProps_perPatient}
\alias{GetProps_perPatient}
\title{GetProps_perPatient}
\usage{
GetProps_perPatient(dat_obj, label, group)
}
\arguments{
\item{dat_obj}{SingleCellExperiment object or subset-able data.frame object}

\item{label}{Character string indicating the labels to calculate proportions for}

\item{group}{Character string indicating the group to calculate proportions within}
}
\value{
data.frame, where rownames are values of `label`, colnames are `group` and values are percentages
}
\description{
Calculates a proportions table for two data annotations. Returns a table where rows are 'label', columns are 'group', and values are percentages
}
