% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunc_R.R
\name{GetShapeLegend}
\alias{GetShapeLegend}
\title{GetShapeLegend}
\usage{
GetShapeLegend(
  shape_dict,
  lab_str,
  title_pos = "top",
  n_row = 1,
  by_row = TRUE,
  rl = 1,
  leg_size = 4,
  buffer = margin(0, 0, 0, 0),
  ...
)
}
\arguments{
\item{shape_dict}{A character vector indicating the legend names for each shape. If named vector, values should be shape numbers}

\item{lab_str}{A character string indicating legend title}

\item{title_pos}{A character string indicating legend position c("top", "bottom")}

\item{n_row}{Integer, number of rows in legend}

\item{by_row}{TRUE or FALSE}

\item{rl}{Relative size}

\item{leg_size}{Integer size of legend keys}

\item{buffer}{Legend margin i.e. margin(t,r,b,l)}

\item{...}{}
}
\value{
A gtable object holding just the legend (from cowplot::get_legend())
}
\description{
Returns a gtable containing a custom colour legend based on shape_dict
}
\examples{

GetShapeLegend(setNames(c(19, 15), c("A", "B")), "Sample")
GetShapeLegend(dge_x$samples$Sample, "Sample")

}
