% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colours.R
\name{getShapes}
\alias{getShapes}
\title{getShapes}
\usage{
getShapes(vec)
}
\arguments{
\item{vec}{A character vector of names}
}
\value{
Named vector of pch symbols to use in plotting
}
\description{
Returns shape palette for names in vec
}
