% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colours.R
\name{GetViridisCol}
\alias{GetViridisCol}
\title{GetViridisCol}
\usage{
GetViridisCol(get = "yellow")
}
\arguments{
\item{get}{Character, if not yellow will return purple value of viridis cmap}
}
\value{
Character, hex colour
}
\description{
Helper function to retrieve end colours of viridis cmap as hex values
}
