% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colours.R
\name{gg_colour_hue}
\alias{gg_colour_hue}
\title{Colour hues}
\usage{
gg_colour_hue(n, lum = 40, chrom = 120)
}
\arguments{
\item{n}{Integer indicating number of colours to return}

\item{lum}{Integer indicating luminance between 0 and 100}

\item{chrom}{Integer indicating chroma of colour}
}
\value{
A character vector of hex colours
}
\description{
Get n hex colours
}
