% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPlot.R
\name{ggDIMRED}
\alias{ggDIMRED}
\title{Create ggplot object using SingleCellExperiment reduceDims}
\usage{
ggDIMRED(
  sce_obj,
  dimred,
  colour_by,
  col_pal = NULL,
  ordering = NULL,
  assay_plot = "logcounts",
  exprs_quant = 0.95,
  npc_units = 0.1,
  point_size = 1.2,
  point_alpha = 0.9
)
}
\arguments{
\item{sce_obj}{SingleCellExperiment object}

\item{dimred}{Character, reduceDim in `sce_obj` to plot}

\item{colour_by}{Character, colData in `sce_obj` to colour by}

\item{col_pal}{Named character vector containing colours for each value in `colour_by`}

\item{ordering}{Character vector of preferred plot ordering of `colour_by` values. Last element in `ordering` will be on top of plot.}

\item{assay_plot}{Character, the assay name to obtain expression values from if `colour_by` is a gene name}

\item{exprs_quant}{Numeric quantile (0,1) for the colour scale if `colour_by` is continuous}

\item{npc_units}{Numeric legend.key.height in npc units if `colour_by` is continuous}

\item{point_size}{Size of points to be passed to ggplot}

\item{point_alpha}{Alpha of points to be passed to ggplot}
}
\value{
A ggplot object
}
\description{
Create ggplot object using SingleCellExperiment reduceDims
}
