% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunc_R.R
\name{ggPCA}
\alias{ggPCA}
\title{ggPCA}
\usage{
ggPCA(
  dat_x,
  annot,
  col_str,
  col_pal = NULL,
  shape_str = NULL,
  shape_pal = NULL,
  isLog = TRUE,
  centre_scale = TRUE,
  point_size = 4,
  k = 2,
  k_vec = NULL
)
}
\arguments{
\item{dat_x}{A subsettable data.frame or matrix object}

\item{annot}{A data.frame object containing sample annotations, including col_str}

\item{col_str}{Character string indicating column of annot to colour points by}

\item{col_pal}{Named character vector to use as colour palette, where values are hex colours and names match col_str}

\item{shape_str}{Character string indicating column of annot to shape points by}

\item{shape_pal}{Named character vector indicating shape legend, where names match shape_str}

\item{isLog}{Defaults to TRUE, set to FALSE if dat_x is not log-transformed to scale prior to PCA calculation}

\item{centre_scale}{TRUE or FALSE should data be centred by scale() prior to PCA calculation?}

\item{point_size}{Integer indicating point size  to provide ggplot}

\item{k}{Defaults to 2, indicates how many principle components to plot}

\item{k_vec}{Vector indicating which principle components to plot i.e. c(1,3) for k=3}
}
\value{
ggplot object
}
\description{
Calculates PCA using calc_pca plots using ggplot
}
\examples{
i.e.
ggPCA(dge_obj$logCPM, Fig4_dge$samples, col_str="Treatment", shape_str = "Patient",isLog=TRUE, k=3)

Or to plot specific PCs
ggPCA(dge_obj$logCPM, Fig4_dge$samples, col_str="Treatment", shape_str = "Patient",isLog=TRUE, k=3, k_vec=c(1,3))

}
