% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleHeatmap.R
\name{MakeAnnot}
\alias{MakeAnnot}
\title{MakeAnnot}
\usage{
MakeAnnot(annot_df, annot_cols, leg_lst, what = "column")
}
\arguments{
\item{annot_df}{Annotation dataframe, rownames are heatmap rownames}

\item{annot_cols}{Named character vector of hex colours for annotations in `annot_df`}

\item{leg_lst}{List of legend params}

\item{what}{Character describing whether `annot_df` contains row or column annotations}
}
\value{
A HeatmapAnnotation-class object
}
\description{
Wrapper function to make heatmap annotation for ComplexHeatmap
}
