% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey.R
\name{MakeSankey3d}
\alias{MakeSankey3d}
\title{MakeSankey3d}
\usage{
MakeSankey3d(ListOfLabels_1, ListOfLabels_2, ListOfLabels_3)
}
\arguments{
\item{ListOfLabels_1}{Character vector of labels}

\item{ListOfLabels_2}{Character vector of labels}

\item{ListOfLabels_3}{Character vector of labels}
}
\value{
A sankeyNetwork() plot
}
\description{
MakeSankey3d
}
