% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunc_R.R
\name{MDS2gg}
\alias{MDS2gg}
\title{MDS2gg}
\usage{
MDS2gg(DGE_x, col_str, col_pal = NULL, shape_str = NULL, shape_pal = NULL)
}
\arguments{
\item{DGE_x}{DGEList object}

\item{col_str}{Character string indicating column of DGE_x$samples to colour points by}

\item{col_pal}{Named character vector to be used as colour palette, where values are hex colours and names match values of col_str}

\item{shape_str}{(Optional) Character string indicating column of DGE_x$samples to shape points by}

\item{shape_pal}{(Optional) Named character vector to indicate shape legend, where names match values of shape_str}
}
\value{
A ggplot object
}
\description{
Converts Limma's MDS into ggplot, input should be DGEList object (edgeR)
}
