% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunc_R.R
\name{plotPCs}
\alias{plotPCs}
\title{plotPCs}
\usage{
plotPCs(
  k_x,
  k_y,
  dat_merge,
  point_size,
  shape_str,
  col_title,
  col_pal,
  shape_pal
)
}
\arguments{
\item{k_x}{Principle component to be plot on x-axis}

\item{k_y}{Principle component to be plot on y-axis}

\item{dat_merge}{A data.frame object from ggPCA() containing PCs and sample annotations}

\item{point_size}{Integer indicating point size  to provide ggplot}

\item{shape_str}{Character string indicating column of annot to shape points by}

\item{col_title}{Character string indicating column of annot to colour points by}

\item{col_pal}{Named character vector to use as colour palette, where values are hex colours and names match col_str}

\item{shape_pal}{Named character vector indicating shape legend, where names match shape_str}
}
\value{
ggplot object
}
\description{
Helper function for ggPCA() that plots individual panels
}
