% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPlot.R
\name{PropBar}
\alias{PropBar}
\title{PropBar}
\usage{
PropBar(
  sce_obj,
  label_str,
  group_str = NULL,
  colour_dict = NULL,
  label_ordering = NULL,
  group_ordering = NULL,
  facet_str = NULL,
  facet_ordering = NULL,
  tick_gap = 2000,
  bar_labels = TRUE,
  WIDTH = 1800,
  HEIGHT = 2000,
  RES = 220,
  rl = 1.1,
  ...
)
}
\arguments{
\item{sce_obj}{SingleCellExperiment object}

\item{label_str}{Character string indicating column of `colData(sce_obj)` to fill proportions of bar}

\item{group_str}{Character string indicating column of `colData(sce_obj)` to group labels. Will have one bar for each of `sce_obj$group_str `}

\item{colour_dict}{Named character vector of hex values, names should contain all of `unique(sce_obj$label_str)`}

\item{label_ordering}{Character vector indicating desired ordering of `label_str`}

\item{group_ordering}{Character vector indicating desired ordering of `group_str`}

\item{facet_str}{Character string indicating column of `colData(sce_obj)` to facet the plot by}

\item{facet_ordering}{Character vector indicating desired ordering of facets (`facet_str`)}

\item{tick_gap}{Numeric spacing between ticks of cell number subplot}

\item{bar_labels}{TRUE or FALSE, whether or not to plot bar labels}

\item{WIDTH}{Width provided to png()}

\item{HEIGHT}{Height provided to png()}

\item{RES}{Res provided to png()}

\item{rl}{Relative text size for legend text and title}
}
\value{
gtable object from gridExtra::grid.arrange()
}
\description{
Plots stacked barchart/s coloured by proportions of label_str in each sample (group_str)
and includes peripheral barcharts showing total cell number per sample.
Calculates proportions with GetProps_perPatient()
If group_str is NULL, PropBar_1D() is called
}
