% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPlot.R
\name{PropBar_1D}
\alias{PropBar_1D}
\title{PropBar_1D}
\usage{
PropBar_1D(sce_obj, label_str, colour_dict, SAVE_PATH = NULL, rl = 1.1)
}
\arguments{
\item{sce_obj}{SingleCellExperiment object}

\item{label_str}{Character string indicating column of colData(sce_obj) to fill proportions of bar}

\item{colour_dict}{Named character vector of hex values, names should contain all of unique(sce_obj$label_str)}

\item{SAVE_PATH}{Optional path to save figure}

\item{rl}{Relative text size for legend text and title}
}
\value{
A ggplot object
}
\description{
Plot a single proportional bar for one label
}
