% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPlot.R
\name{scHeatmap}
\alias{scHeatmap}
\title{scHeatmap}
\usage{
scHeatmap(
  sce_obj,
  group_str,
  genes_to_subset,
  exprs_val = "TPM",
  log = TRUE,
  save_path = NULL,
  scale_str = "genes",
  cluster_genes = FALSE,
  cluster_samples = FALSE,
  groups_to_plot = NULL,
  min_cells = 100,
  lower_thresh = 0.05,
  upper_thresh = 0.95,
  rl = 1.1,
  label_side = "bottom"
)
}
\arguments{
\item{sce_obj}{SingleCellExperiment object}

\item{group_str}{Character string indicating column group of `colData(sce_obj)` to average within}

\item{genes_to_subset}{Character vector of gene names, should be present in `rownames(sce_obj)`}

\item{exprs_val}{Character string indicating expression values to use for averaging, one of c("TPM", "counts","logcounts")}

\item{log}{TRUE or FALSE, whether or not to log `exprs_val`}

\item{save_path}{Optional path to save figure, if NULL function returns figure}

\item{scale_str}{Character string indicating how to scale heatmap, one of c("genes","samples", "both")}

\item{cluster_genes}{TRUE or FALSE, whether or not to cluster genes (rows)}

\item{cluster_samples}{TRUE or FALSE, whether or not to cluster samples (columns)}

\item{groups_to_plot}{Character vector indicating a subset of `group_str` to plot}

\item{min_cells}{Minimum number of cells required in a group to plot that group if `groups_to_plot` is NULL}

\item{lower_thresh}{Lower quantile of expression values to trim heatmap, default is 0.05}

\item{upper_thresh}{Upper quantile of expression values to scale colour bar, default is 0.95}

\item{rl}{Numeric, relative text size, default is 1.1}

\item{label_side}{One of c("bottom", "top") to indicate if `group_str` labels should be below or above heatmap, defaults to "bottom"}
}
\value{
A ggplot object
}
\description{
Plots heatmap of average per-group expression of genes in genes_to_subset.
Groups (group_str) are columns of heatmap and genes (genes_to_subset) are rows.
}
