% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scWrangle.R
\name{Seurat2SCE}
\alias{Seurat2SCE}
\title{Seurat2SCE}
\usage{
Seurat2SCE(seurat_obj, dat_count = "raw")
}
\arguments{
\item{seurat_obj}{Input Seurat object}

\item{dat_count}{Character string indicating which Seurat assay to use}
}
\value{
SingleCellExperiment object
}
\description{
Helper function for converting seurat objects to SingleCellExperiment
}
