% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colours.R
\name{ShowColPal}
\alias{ShowColPal}
\title{ShowColPal}
\usage{
ShowColPal(colour_vec, name_str = NULL, title_mar = -20)
}
\arguments{
\item{colour_vec}{Character vector containing hex colours}

\item{name_str}{Name of colour palette, if not NULL a title is added}

\item{title_mar}{Integer value for margin between plot and title i.e. margin(0,0,title_mar,0)}
}
\value{
Barplot, displaying colours in `colour_vec`
}
\description{
Plot input colour palette as ggplot object, uses geom_bar
}
\examples{

## Show what a palette looks like
ShowColPal(hcl.colors(100, palette = "Mako"))

## Plot palette options for n=4
lapply(GetColPal_options(4), ShowColPal)

## Pick between sequential colour palettes
plot_lst <- lapply(hcl.pals("sequential"),
function(x){ShowColPal(hcl.colors(10, palette = x), name_str=x)})
cowplot::plot_grid(plotlist = plot_lst[1:20], ncol = 2)

}
