% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunc_R.R
\name{theme_blank}
\alias{theme_blank}
\title{theme_blank}
\usage{
theme_blank(legend = "none")
}
\arguments{
\item{legend}{Legend position, defaults to "none"}
}
\value{
A ggplot theme
}
\description{
Add a blank theme to ggplot
}
