% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunc_R.R
\name{UpsetPlot_wBars}
\alias{UpsetPlot_wBars}
\title{UpsetPlot_wBars}
\usage{
UpsetPlot_wBars(DATA, col_names, bar_col, name_str = "", col_pal = NULL)
}
\arguments{
\item{DATA}{SingleCellExperiment object or data.frame, data object containing continuous variables that can be subset. Rows are observations, columns are variables.}

\item{col_names}{Character vector of column names, often gene names}

\item{bar_col}{Character string of column in `DATA` used to plot proportional bars, should be a discrete/character variable}

\item{name_str}{Character string, a general name for the variables in `col_names`}

\item{col_pal}{Named character vector, values should be hex colour values and names should correspond to `DATA$bar_col`}
}
\value{
ComplexUpset::upset() object
}
\description{
Wrapper function for ComplexUpset::upset() to plot binarised continuous variables, includes proportional bar plots
}
